requirements.txt
mlx-vlm>=0.0.9
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
